fix(tui): close the exit/output race and the Windows build
Four CI failures, three distinct causes.

**A real race in the hosted loop (test ubuntu-latest).** The reaper and
the reader are separate threads, so `HostEvent::Exited` can arrive before
the reader has delivered the child's final bytes. `hosted_loop` returned
on `Exited` after a `try_recv` drain, which only picks up what is already
queued — so a fast-exiting child could lose its last frame, or on Linux
its entire output. The test asserted the same wrong contract and passed
on macOS purely because the ordering happened to favour output first.

The channel closing is the real end of the session: it means every sender
dropped, so the reader reached EOF too. The loop now records the status on
`Exited` and keeps draining until `None`, bounded by a 500ms deadline for
the case where a descendant holds the pty open and EOF never comes. The
test drains to channel close for the same reason.

**Windows build (clippy + test windows-latest).** `handle_key` and `draw`
referenced a `#[cfg(unix)]` enum without being gated themselves. The
underlying problem was eight scattered per-item gates, which is a
structure that invites exactly this: the interactive machinery moves to
`tui/watch.rs`, gated once at the module declaration, so unix-only is a
property of the file rather than something to remember per item. A
`cfg!(unix)` runtime check in `run_launch` had the same flaw — it cannot
gate a call to `exec_hosted`, which only exists under the gate — and is
now a compile-time branch.

Verified by compiling the non-unix arm locally: flipping only the gates
this branch introduced (`cfg(unix)` -> `cfg(any())`) and building on
macOS. Cross-compiling to a Windows target is not possible here, since
`ring` and `aws-lc-sys` build scripts need a target C toolchain.

**Doc build.** Two unresolved intra-doc links under `RUSTDOCFLAGS=-D
warnings`. `RoutingOverlay` had been broken all along but was invisible
while `ChildLaunch` was private — making it public is what made rustdoc
check it.

Verified with CI's exact commands (`clippy --workspace --all-features
--tests -- -D warnings`, `RUSTDOCFLAGS=-D warnings cargo doc`, `nextest
run --workspace --all-features`) plus 15 consecutive runs of the pty
tests under CPU load to confirm the race is closed rather than hidden.
